+++ src/app.rs
@@ -4,12 +4,14 @@ use handle_rep::HandledReputation;
 use message_data::MessageData;
 use std::future::Future;
 use teloxide::prelude::*;
+use crate::app::reputation_message::ReputationMessage;
 use crate::services::bootstrap;
 
 pub mod handle_cmd;
 pub mod handle_rep;
 pub mod handle_user;
 pub mod message_data;
+pub mod reputation_message;
 
 pub fn init() -> impl Future {
     bootstrap::start();
@@ -18,7 +20,8 @@ pub fn init() -> impl Future {
     teloxide::repl(bot, |message: Message, bot: AutoSend<Bot>| async move {
         handle_user::save_user(&message);
         let data = MessageData::get_data(&message);
-
+        let reputation = ReputationMessage::new(&message);
+        let a = message.reply_to_message();
         let (is_command, command_message) = handle_cmd::execute(&message);
         if is_command {
             let command_message = command_message.unwrap_or("Unknown command error".to_string());
@@ -27,7 +30,7 @@ pub fn init() -> impl Future {
             return respond(());
         }
 
-        let result = HandledReputation::handle_rep(&data);
+        let result = HandledReputation::handle_rep(&reputation);
         match result {
             Some(_handled_reputation) => {
                 MessageSender::new(data.get_chat_id(), _handled_reputation.parse())