+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Simple migrations system
+### Changed
+- Reputation triggers
+  - change structure
+  - a
+
 ## [0.2.0] - 2022-06-13
 
 ### Added